refactor:moved integration test files to separate directory
