+++ tax/src/test/java/com/yas/tax/repository/TaxClassRepoDataJPATest.java
@@ -2,6 +2,7 @@ package com.yas.tax.repository;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import com.yas.tax.config.DatabaseAutoConfig;
 import com.yas.tax.model.TaxClass;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -9,23 +10,19 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.FilterType;
-import org.springframework.data.auditing.AuditingHandler;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @DataJpaTest(properties = {
     "spring.datasource.url=jdbc:h2:mem:testdb",
     "spring.jpa.hibernate.ddl-auto=create-drop"
-},includeFilters = @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE,
-    classes = AuditingHandler.class),
-    excludeFilters = @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE,
+}, excludeFilters = @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE,
     classes = TaxRateRepository.class))
 @ExtendWith(SpringExtension.class)
-public class TaxClassRepositoryDataJPATest {
-    @MockBean
-    private AuditingHandler auditingHandler;
+@Import(DatabaseAutoConfig.class)
+public class TaxClassRepoDataJPATest {
     @Autowired
     private TaxClassRepository taxClassRepository;
